==> notebooks/2. feature_engineering.ipynb <==
import pandas as pd
import numpy as np
import pickle

import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.feature_selection import VarianceThreshold
from sklearn.impute import SimpleImputer
from sklearn.compose import make_column_transformer
from sklearn.preprocessing import OneHotEncoder, StandardScaler


pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 100)

import sys
import os

PROJECT_DIRECTORY = os.path.join(os.getcwd(), '..')
sys.path.append(PROJECT_DIRECTORY)

from src.utils import *

abt_treino = pd.read_parquet(PROJECT_DIRECTORY + '/data/processed/abt_treino.parquet')

# Metadados da Base

metadados = gerar_metadados(abt_treino,['id','safra'],'y')
metadados

# Remoção de variáveis com altas taxas de nulos

# Limiar para corte de percentual de nulos
THRESHOLD_PC_NULOS = 60

metadados_nulos_filtrados = metadados[metadados['PC_NULOS'] < THRESHOLD_PC_NULOS]

print(f'Quantidade de variáveis iniciais: {metadados.shape[0]}')
print(f'Quantidade de variáveis após filtragem dos nulos: {metadados_nulos_filtrados.shape[0]}')

# Identificando variáveis categóricas

TARGET = ['y']
AUXILIARES = ['id','safra']

# Limiar para classificação como categórica
THRESHOLD_CARDINALIDADE = 25

NUM_VARS = metadados_nulos_filtrados[(metadados_nulos_filtrados['CARDINALIDADE'] > THRESHOLD_CARDINALIDADE) & (metadados_nulos_filtrados['USO_FEATURE'] == 'Explicativa')]['FEATURE'].to_list()
CAT_VARS = metadados_nulos_filtrados[(metadados_nulos_filtrados['CARDINALIDADE'] <= THRESHOLD_CARDINALIDADE) & (metadados_nulos_filtrados['USO_FEATURE'] == 'Explicativa')]['FEATURE'].to_list()

print(f'Quantidade de Variáveis Númericas: {len(NUM_VARS)}')
print(f'Quantidade de variáveis categóricas: {len(CAT_VARS)}')

# Remoção de variáveis altamente correlacionadas - Variáveis Númericas 

THRESHOLD_CORRELACAO = 0.9

pares_corr, colunas_remover = tratar_multicolinearidade(abt_treino[NUM_VARS], limiar = THRESHOLD_CORRELACAO)

print("Pares com alta correlação:")
print(pares_corr)

print("\nColunas sugeridas para remoção (baseado em nulos e cardinalidade):")
print(colunas_remover)

NUM_VARS_V2 = [x for x in NUM_VARS if x not in colunas_remover]
print(f'Quantidade de Variáveis Númericas: {len(NUM_VARS_V2)}')

# Remoção de variáveis com baixa variância - Variáveis Númericas 

# Remoção de variaveis com um limiar de variância abaixo de 0,05
THRESHOLD_VARIANCIA = 0.05

selector = VarianceThreshold(threshold=THRESHOLD_VARIANCIA)
df_filtrado = abt_treino[NUM_VARS_V2]
df_filtrado = df_filtrado.loc[:, selector.fit(df_filtrado).get_support()]

print(f'Quantidade de Variáveis Númericas: {df_filtrado.shape[1]}')

# Imputação dos nulos - Variáveis Númericas 

imputer = SimpleImputer(strategy='mean')
abt_treino[NUM_VARS_V2] = imputer.fit_transform(abt_treino[NUM_VARS_V2])

IMPUTER_PATH = PROJECT_DIRECTORY + "/models/artefacts/num_imputer.pickle"

# Salvar o imputer treinado com pickle
with open(IMPUTER_PATH, 'wb') as f:
    pickle.dump(imputer, f)

# Normalização - Variáveis Númericas

scaler = StandardScaler()
abt_treino[NUM_VARS_V2] = scaler.fit_transform(abt_treino[NUM_VARS_V2])

SCALER_PATH = PROJECT_DIRECTORY + "/models/artefacts/num_scaler.pickle"

# Salvar o imputer treinado com pickle
with open(SCALER_PATH, 'wb') as f:
    pickle.dump(scaler, f)

# Colunas duplicadas - Variáveis Categóricas

colunas_duplicadas = encontrar_colunas_duplicadas(abt_treino)
print(colunas_duplicadas)

# Remoção de variáveis altamente correlacionadas - Variáveis Categóricas 

THRESHOLD_CORRELACAO_CAT = 0.9

pares_corr, colunas_remover = tratar_multicolinearidade_cat(abt_treino,CAT_VARS, limiar=THRESHOLD_CORRELACAO_CAT)

print("Pares com alta correlação:")
print(pares_corr)

print("\nColunas sugeridas para remoção (baseado em nulos e cardinalidade):")
print(colunas_remover)

CAT_VARS_V2 = [x for x in CAT_VARS if x not in colunas_remover]
print(f'Quantidade de Variáveis Categóricas: {len(CAT_VARS_V2)}')

# Imputação dos nulos - Variáveis Categóricas 

cat = abt_treino[[*CAT_VARS_V2,*TARGET]]
metadados_nulos_filtrados[metadados_nulos_filtrados['FEATURE'].isin(CAT_VARS_V2)]

THRESHOLD_HN = 30

CAT_VARS_V2_HN = [col for col in 
              metadados_nulos_filtrados[(metadados_nulos_filtrados['FEATURE'].isin(CAT_VARS_V2)) &
                                        (metadados_nulos_filtrados['PC_NULOS'] >= THRESHOLD_HN)
                                        ]['FEATURE']]

CAT_VARS_V2_LN = [coluna for coluna in CAT_VARS_V2 if coluna not in CAT_VARS_V2_HN]

INPUTS_LN = {col: abt_treino[col].mode(dropna=True)[0] for col in CAT_VARS_V2_LN}
INPUTS_HN = {col: 9999 for col in abt_treino[CAT_VARS_V2_HN].columns}

INPUTS_CAT = {**INPUTS_LN,**INPUTS_HN}

# Aplicar a imputação nas variáveis
for col, valor in INPUTS_CAT.items():
    abt_treino[col] = abt_treino[col].fillna(valor)

# Salvando os valores para tratamento de nulos das variáveis categóricas
IMPUTER_CAT_PATH = PROJECT_DIRECTORY + "/models/artefacts/cat_imputer.pickle"

with open(IMPUTER_CAT_PATH, 'wb') as f:
    pickle.dump(INPUTS_CAT, f)

# Dumming - Variáveis categóricas

one_hot_enc = make_column_transformer(
    (OneHotEncoder(handle_unknown = 'ignore',sparse_output=False),
    CAT_VARS_V2),
    remainder='drop')

dados_one_hot = one_hot_enc.fit_transform(abt_treino)
dados_one_hot = pd.DataFrame(dados_one_hot,
                     columns=one_hot_enc.get_feature_names_out(),
                     index=abt_treino.index)
dados_one_hot

# Salvando os valores para tratamento de nulos das variáveis categóricas
ONE_HOT_PATH = PROJECT_DIRECTORY + "/models/artefacts/cat_encoder.pickle"

with open(ONE_HOT_PATH, 'wb') as f:
    pickle.dump(one_hot_enc, f)

# Salvando Colunas Numéricas e Categóricas

NUM_VARS_PATH = PROJECT_DIRECTORY + '/models/artefacts/num_vars.pickle'
CAT_VARS_PATH = PROJECT_DIRECTORY + '/models/artefacts/cat_vars.pickle'
TARGET_VARS_PATH = PROJECT_DIRECTORY + '/models/artefacts/target_vars.pickle'
AUX_VARS_PATH = PROJECT_DIRECTORY + '/models/artefacts/aux_vars.pickle'

with open(NUM_VARS_PATH,'wb') as f:
    pickle.dump(NUM_VARS_V2,f)

with open(CAT_VARS_PATH,'wb') as f:
    pickle.dump(CAT_VARS_V2,f)  

with open(TARGET_VARS_PATH,'wb') as f:
    pickle.dump(TARGET,f)  

with open(AUX_VARS_PATH,'wb') as f:
    pickle.dump(AUXILIARES,f)  

# Função de Tratamento Final

def tratamento_completo(df, aux_vars_path,target_var_path,num_vars_path,cat_vars_path,cat_imputer_path,num_imputer_path,cat_encoder_path,num_scaler_path):
    """
    Aplica um pipeline completo de tratamento de dados a um DataFrame, incluindo seleção de variáveis,
    imputação de nulos, codificação de variáveis categóricas e normalização de variáveis numéricas.

    Parâmetros
    ----------
    df : pandas.DataFrame
        DataFrame de entrada com os dados brutos.

    aux_vars_path : str
        Caminho para o arquivo `.pickle` contendo a lista de variáveis auxiliares.

    target_var_path : str
        Caminho para o arquivo `.pickle` contendo a lista de variáveis alvo.

    num_vars_path : str
        Caminho para o arquivo `.pickle` contendo a lista de variáveis numéricas.

    cat_vars_path : str
        Caminho para o arquivo `.pickle` contendo a lista de variáveis categóricas.

    cat_imputer_path : str
        Caminho para o arquivo `.pickle` contendo o dicionário de imputação de variáveis categóricas.

    num_imputer_path : str
        Caminho para o arquivo `.pickle` contendo o imputador de variáveis numéricas.

    cat_encoder_path : str
        Caminho para o arquivo `.pickle` contendo o codificador para variáveis categóricas.

    num_scaler_path : str
        Caminho para o arquivo `.pickle` contendo o normalizador para variáveis numéricas.

    Retorna
    -------
    pandas.DataFrame
        DataFrame tratado com variáveis imputadas, codificadas e normalizadas, pronto para modelagem.
    """

    with open(aux_vars_path,'rb') as f:
        aux_vars = pickle.load(f)

    with open(target_var_path,'rb') as f:
        target_vars = pickle.load(f)
    
    with open(num_vars_path,'rb') as f:
        num_vars = pickle.load(f)

    with open(cat_vars_path,'rb') as f:
        cat_vars = pickle.load(f)

    with open(cat_imputer_path,'rb') as f:
        cat_imputer = pickle.load(f)

    with open(num_imputer_path,'rb') as f:
        num_imputer = pickle.load(f)

    with open(cat_encoder_path,'rb') as f:
        cat_encoder = pickle.load(f)

    with open(num_scaler_path,'rb') as f:
        num_scaler = pickle.load(f)  

    # Selecionando Colunas 
    df = df.reindex(columns=[*aux_vars,
                             *target_vars,
                             *num_vars,
                             *cat_vars])
    
    
    # Imputando Nulos
    df[num_vars] = num_imputer.transform(df[num_vars])
    
    for col, valor in cat_imputer.items():
        df[col] = df[col].fillna(valor)
   
    # Dummy encoding
    dados_one_hot = cat_encoder.transform(df)
    dados_one_hot = pd.DataFrame(dados_one_hot,
                     columns=cat_encoder.get_feature_names_out(),
                     index=df.index)
    
    df = pd.concat([df,dados_one_hot],axis=1)
    df = df.drop(columns=cat_vars)

    # Normalização
    df[num_vars] = num_scaler.transform(df[num_vars])

    
    return df

df_teste = pd.read_parquet(PROJECT_DIRECTORY + '/data/processed/abt_treino.parquet')

aux_vars_path    = PROJECT_DIRECTORY + '/models/artefacts/aux_vars.pickle'
target_var_path  = PROJECT_DIRECTORY + '/models/artefacts/target_vars.pickle'
num_vars_path    = PROJECT_DIRECTORY + '/models/artefacts/num_vars.pickle'
cat_vars_path    = PROJECT_DIRECTORY + '/models/artefacts/cat_vars.pickle'

cat_imputer_path = PROJECT_DIRECTORY + '/models/artefacts/cat_imputer.pickle'
num_imputer_path = PROJECT_DIRECTORY + '/models/artefacts/num_imputer.pickle'

cat_encoder_path = PROJECT_DIRECTORY + '/models/artefacts/cat_encoder.pickle'
num_scaler_path  = PROJECT_DIRECTORY + '/models/artefacts/num_scaler.pickle'


df_teste = tratamento_completo(df = df_teste,
                               aux_vars_path    = aux_vars_path,
                               target_var_path  = target_var_path,
                               num_vars_path    = num_vars_path,
                               cat_vars_path    = cat_vars_path,
                               cat_imputer_path = cat_imputer_path,
                               num_imputer_path = num_imputer_path,
                               cat_encoder_path = cat_encoder_path,
                               num_scaler_path  = num_scaler_path)

df_teste.head()
