==> notebooks/4. model_training.ipynb <==
import pandas as pd

from sklearn.model_selection import StratifiedKFold

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier

import sys
import os

PROJECT_DIRECTORY = os.path.join(os.getcwd(), '..')
sys.path.append(PROJECT_DIRECTORY)

from src.features_processing import *
from src.model_evaluation import *

# Carregando Artefatos

AUX_VAR_PATH     = PROJECT_DIRECTORY + '/models/artefacts/aux_vars.pickle'
TARGET_VARS_PATH = PROJECT_DIRECTORY + '/models/artefacts/target_vars.pickle'
NUM_VARS_PATH    = PROJECT_DIRECTORY + '/models/artefacts/num_vars.pickle'
CAT_VARS_PATH    = PROJECT_DIRECTORY + '/models/artefacts/cat_vars.pickle'

CAT_IMPUTER_PATH = PROJECT_DIRECTORY + '/models/artefacts/cat_imputer.pickle'
NUM_IMPUTER_PATH = PROJECT_DIRECTORY + '/models/artefacts/num_imputer.pickle'

CAT_ENCONDER_PATH = PROJECT_DIRECTORY + '/models/artefacts/cat_encoder.pickle'
NUM_SCALER_PATH   = PROJECT_DIRECTORY + '/models/artefacts/num_scaler.pickle'

SELECTED_VARS_PATH = PROJECT_DIRECTORY + '/models/artefacts/selected_vars.pickle'


with open(AUX_VAR_PATH,'rb') as f:
    aux_vars = pickle.load(f)

with open(TARGET_VARS_PATH,'rb') as f:
    target_vars = pickle.load(f)

with open(SELECTED_VARS_PATH,'rb') as f:
    selected_vars = pickle.load(f)

# Carregando bases de treino e teste

abt_treino = pd.read_parquet(PROJECT_DIRECTORY + '/data/processed/abt_treino.parquet')
abt_teste = pd.read_parquet(PROJECT_DIRECTORY + '/data/processed/abt_teste.parquet')


abt_treino = tratamento_completo(df = abt_treino,
                                 aux_vars_path    = AUX_VAR_PATH,
                                 target_var_path  = TARGET_VARS_PATH,
                                 num_vars_path    = NUM_VARS_PATH,
                                 cat_vars_path    = CAT_VARS_PATH,
                                 cat_imputer_path = CAT_IMPUTER_PATH,
                                 num_imputer_path = NUM_IMPUTER_PATH,
                                 cat_encoder_path = CAT_ENCONDER_PATH,
                                 num_scaler_path  = NUM_SCALER_PATH)


abt_teste = tratamento_completo(df = abt_teste,
                                 aux_vars_path    = AUX_VAR_PATH,
                                 target_var_path  = TARGET_VARS_PATH,
                                 num_vars_path    = NUM_VARS_PATH,
                                 cat_vars_path    = CAT_VARS_PATH,
                                 cat_imputer_path = CAT_IMPUTER_PATH,
                                 num_imputer_path = NUM_IMPUTER_PATH,
                                 cat_encoder_path = CAT_ENCONDER_PATH,
                                 num_scaler_path  = NUM_SCALER_PATH)

# Preparando variaveis para treino e teste

X_train  = abt_treino[selected_vars]
y_train  = abt_treino[target_vars]

X_test = abt_teste[selected_vars]
y_test = abt_teste[target_vars]

# Definindo Modelos

models = {
    'Logistic Regression': LogisticRegression(max_iter=1000, random_state=42),
    'Random Forest':       RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=42),
    'XGBoost':             XGBClassifier(use_label_encoder=False, eval_metric='logloss', random_state=42),
    'LightGBM':            LGBMClassifier(n_estimators=100, n_jobs=-1, random_state=42)
}

# Cross Validation

# Configura o Stratified K-Fold sobre o treino
n_splits = 5
skf = StratifiedKFold(n_splits=n_splits, shuffle=True, random_state=42)

# Cross-validation dados de treino
cv_results = {name: {'ks':[], 'auc':[], 'gini':[]} for name in models}

for fold, (idx_tr, idx_val) in enumerate(skf.split(X_train, y_train), 1):
    X_tr, X_val = X_train.iloc[idx_tr], X_train.iloc[idx_val]
    y_tr, y_val = y_train.iloc[idx_tr], y_train.iloc[idx_val]
    
    for name, model in models.items():
        model.fit(X_tr, y_tr)
        y_prob = model.predict_proba(X_val)[:,1]
        ks, auc, gini = performance_metrics(y_val, y_prob)
        cv_results[name]['ks'].append(ks)
        cv_results[name]['auc'].append(auc)
        cv_results[name]['gini'].append(gini)

# Agregando as métricas de CV
cv_summary = []
for name, m in cv_results.items():
    cv_summary.append({
        'Modelo':    name,
        'KS Médio':  np.mean(m['ks']),  'KS Desvio':   np.std(m['ks']),
        'AUC Médio': np.mean(m['auc']), 'AUC Desvio':  np.std(m['auc']),
        'Gini Médio':np.mean(m['gini']),'Gini Desvio': np.std(m['gini']),
    })
    
df_cv = pd.DataFrame(cv_summary).sort_values('KS Médio', ascending=False)

# Avaliação final no Teste
test_summary = []
for name, model in models.items():
    model.fit(X_train, y_train) 
    y_prob_test = model.predict_proba(X_test)[:,1]
    ks_t, auc_t, gini_t = performance_metrics(y_test, y_prob_test)
    test_summary.append({
        'Modelo':    name,
        'KS':       ks_t,
        'AUC':      auc_t,
        'Gini':     gini_t
    })
    
df_test = pd.DataFrame(test_summary).sort_values('AUC', ascending=False)

print("=== Resultados Cross-Validation - Treino ===")
display(df_cv)

print("\n=== Resultados Finais - Teste ===")
display(df_test)
